pyproject.toml
[project]
name = "specx-cli"
version = "1.0.0"
description = "SpecX Bot - AI-powered Spec-Driven Development CLI with role-based personas. Forked from GitHub Spec Kit."
requires-python = ">=3.11"
dependencies = [
    "typer",
    "rich",
    "httpx[socks]",
    "platformdirs",
    "readchar",
    "truststore>=0.10.4",
    "pyyaml",
]

[project.scripts]
specx = "specx_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/specx_cli"]

[tool.hatch.build.targets.wheel.force-include]
"templates/chatmodes" = "templates/chatmodes"

